add ps_client
