+++ elasticdl/python/tests/worker_ps_interaction_test.py
@@ -180,7 +180,7 @@ class WorkerPSInteractionTest(unittest.TestCase):
 
         result_dict = {}
         for layer in layers:
-            embedding = worker.pull_embedding_vectors(layer, ids)
+            embedding = worker._ps_client.pull_embedding_vectors(layer, ids)
             result_dict[layer] = embedding
 
         for layer in layers:
